+++ src/gmt_init.c
@@ -10564,7 +10564,7 @@ int gmt_parse_vector (struct GMT_CTRL *GMT, char symbol, char *text, struct GMT_
 	S->v.status = 0;	/* Start with no flags turned on */
 	S->v.v_angle = 30.0f;	S->v.v_norm = -1.0f;	S->v.v_stem = 0.1f;
 	S->v.v_kind[0] = S->v.v_kind[1] = GMT_VEC_ARROW;
-	S->v.v_shape = GMT->current.setting.map_vector_shape;	/* Can be overridden with +h<shape> */
+	S->v.v_shape = (float)GMT->current.setting.map_vector_shape;	/* Can be overridden with +h<shape> */
 	for (k = 0; text[k] && text[k] != '+'; k++);	/* Either find the first plus or run out or chars */
 	strncpy (p, text, k); p[k] = 0;
